+++ apps/commerce-streamer/src/main/java/com/loopers/interfaces/consumer/audit/AuditKafkaListener.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.loopers.config.kafka.KafkaConfig;
 import com.loopers.domain.audit.AuditCommand;
 import com.loopers.domain.audit.AuditService;
-import com.loopers.interfaces.consumer.metrics.LikeEvent;
-import com.loopers.interfaces.consumer.metrics.OrderEvent;
-import com.loopers.interfaces.consumer.metrics.ProductEvent;
+import com.loopers.interfaces.consumer.events.LikeEvent;
+import com.loopers.interfaces.consumer.events.OrderEvent;
+import com.loopers.interfaces.consumer.events.ProductEvent;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
