+++ apps/commerce-streamer/src/main/java/com/loopers/interfaces/consumer/metrics/MetricKafkaConsumer.java
@@ -4,6 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.loopers.application.metrics.MetricCriteria;
 import com.loopers.application.metrics.MetricsFacade;
 import com.loopers.config.kafka.KafkaConfig;
+import com.loopers.interfaces.consumer.events.LikeEvent;
+import com.loopers.interfaces.consumer.events.OrderEvent;
+import com.loopers.interfaces.consumer.events.ProductEvent;
 import java.io.IOException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
